Isolate WebSockets from each other
Now the interface uses a `makeAgent` function to give each websocket client a "fresh" agent. We could also consider re-using Agent but having a separate Chat, or Request, since otherwise the event-handling callbacks that we register on the agent end up being registered for _every_ client on every response event.